Adding RobustOrdinalEncoder
RobustOrdinalEncoder acts like an sklearn OrdinalEncoder, but does not throw an exception when encountering an
unobserved value. Instead, it assigns it to the integer num_values. For example, if during fit we observe two
values 'Cat', 'Dog', but in transform we observe 'Elephant' and 'Horse'. Both will be encoded as 2.
In the inverse transform the mapping of unknown values is converted to None. Notice that this will cause the array
type to be 'object'. This is consistent with the behaviour of sklearn's OneHotEncoder.