feature: dummy feature commit for RobustOrdinalEncoder & add badges to README
This commit is a dummy commit for de2174c279e30b87572ced2bb8e85c8bab30132d.
The empty dummy commit is used to correct the commit message to have the prefix "feature".
That prefix will allow the CICD to trigger the correct version upgrade to follow semantic versioning.
The addition of the "feature" prefix will bump the minor version rather than the patch version.